clean code and manage auth
