+++ back/app/config.py
@@ -7,3 +7,9 @@ IMAGE_PATH = os.path.join(ROOT_DIR, 'data/images')
 class SuperUser:
     USERNAME = os.environ['SUPERUSER_USERNAME']
     PWD = os.environ['SUPERUSER_PWD']
+
+
+# Secret key for JWT encoding and decoding
+SECRET_KEY = os.getenv('SECRET_KEY', "")
+ALGORITHM = os.getenv('ALGORITHM', "HS256")
+ACCESS_TOKEN_EXPIRE_MINUTES = os.getenv('ACCESS_TOKEN_EXPIRE_MINUTES', 30)
