--- back/app/src/entities/models.py
@@ -1,79 +1,0 @@
-import json
-import logging
-
-import bcrypt
-
-from sqlalchemy import Column, Integer, String, TIMESTAMP, func, ForeignKey, Boolean
-
-from app.src.entities.database import Base
-
-logger = logging.getLogger(__name__)
-
-created_date_comment = "Date of the creation of the field"
-updated_date_comment = "Last updated date"
-datetime_format = '%Y-%m-%dT%H:%M:%S:%f'
-
-
-class User(Base):
-    __tablename__ = "users"
-    __table_args__ = {
-        'comment': 'Table containing user information.'
-    }
-    id = Column(Integer, primary_key=True)
-    username = Column(String, nullable=False, unique=True, comment="email or service name")
-    created_at = Column(TIMESTAMP(timezone=True), server_default=func.now(), nullable=False,
-                        comment=created_date_comment)
-    updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now(), onupdate=func.now(), nullable=False,
-                        comment=updated_date_comment)
-
-    hash_password = Column(String, nullable=False)
-
-    @staticmethod
-    def set_password(password):
-        hashed_password = bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt())
-        return hashed_password.decode('utf-8')
-
-    def check_password(self, password):
-        return bcrypt.checkpw(password.encode('utf-8'), self.hash_password.encode('utf-8'))
-
-    def to_dict(self):
-        return {
-            "id": self.id,
-            "username": self.username,
-            "created_at": self.created_at.strftime(datetime_format),
-            "updated_at": self.updated_at.strftime(datetime_format)
-        }
-
-    def to_json(self):
-        return json.dumps(self.to_dict())
-
-
-class OriginalArt(Base):
-    __tablename__ = "original_arts"
-    __table_args__ = {
-        'comment': 'Table containing all original arts and their information.'
-    }
-    id = Column(Integer, primary_key=True, nullable=False, comment="ID generated when creating element")
-
-    filepath = Column(String, nullable=False, comment="Relative filepath of the image")
-    explanation = Column(String, nullable=False, comment="Text which explained the original art")
-    author = Column(String, nullable=False, comment="Author information")
-    gen_description = Column(String, nullable=False, comment="Description generated by the LLM")
-    created_at = Column(TIMESTAMP(timezone=True), server_default=func.now(), nullable=False,
-                        comment=created_date_comment)
-    updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now(), onupdate=func.now(), nullable=False,
-                        comment=updated_date_comment)
-
-    def to_dict(self):
-        return {
-            "id": self.id,
-            "filepath": self.filepath,
-            "explanation": self.explanation,
-            "author": self.author,
-            "gen_description": self.gen_description,
-            "created_at": self.created_at.strftime(datetime_format),
-            "updated_at": self.updated_at.strftime(datetime_format)
-        }
-
-    def to_json(self):
-        return json.dumps(self.to_dict())
