+++ back/app/src/entities/schemas/user.py
@@ -18,3 +18,7 @@ class UserDB(UserBase):
 
 class UserCreate(UserBase):
     password: str
+
+
+class DecodedUser(UserBase):
+    id: int
