+++ back/app/src/services/tables/user.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Session
 
-from app.src.entities.models import User as ModelUser
+from app.src.entities.models.user import User as ModelUser
 from app.src.entities.schemas.user import UserCreate
 from app.src.services.tables.crud_abstract import CRUDAbstractService
 
@@ -12,7 +12,6 @@ class UserService(CRUDAbstractService):
 
     @staticmethod
     def create_model(user: UserCreate) -> ModelUser:
-
         model_user = ModelUser(
             username=user.username,
             hash_password=ModelUser.set_password(user.password),
@@ -20,3 +19,6 @@ class UserService(CRUDAbstractService):
 
         return model_user
 
+    def get_user_by_username(self, username: str) -> ModelUser | None:
+        model_data = self.session.query(self.table).filter(self.table.username == username).first()
+        return model_data
