+++ back/app/src/utils/const.py
@@ -1,0 +1,9 @@
+# DB part
+CREATED_DATE_COMMENT = "Date of the creation of the field"
+UPDATED_DATE_COMMENT = "Last updated date"
+DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%S:%f'
+
+
+# EXCEPTIONS part
+INCORRECT_CREDENTIALS = "Incorrect username or password"
+USER_NOT_FOUND = "User is not registered in the database"
